Move the sanitizeText function to the chatFormat directive
