+++ src/app/chat/chat.component.ts
@@ -82,18 +82,4 @@ export class ChatComponent implements OnInit {
       }
     );
   }
-
-sanitizeText(text: string): string {
-  // Replace newline characters with <br> tags
-  const textWithLineBreaks = text.replace(/\n/g, '<br>');
-
-  // Replace code block markers with <code> tags
-  const codeBlockStart = /```/g;
-  const codeBlockEnd = /```/g;
-  const textWithCodeBlocks = textWithLineBreaks
-    .replace(codeBlockStart, '<code>')
-    .replace(codeBlockEnd, '</code>');
-
-  return textWithCodeBlocks;
-}
 }
